+++ plone/app/upgrade/v51/final.py
@@ -52,3 +52,11 @@ def remove_highlightsearchterms(context):
     resources = registry.records[record]
     if u'jquery-highlightsearchterms' in resources.value:
         resources.value.remove(u'jquery-highlightsearchterms')
+
+
+def remove_old_PAE_rescources(context):  # noqa
+    """FORCE remove old p.a.event resources"""
+    registry = getUtility(IRegistry)
+    plone_legacy = registry.records['plone.bundles/plone-legacy.resources']
+    plone_legacy.value.remove('resource-plone-app-event-event-js')
+    plone_legacy.value.remove('resource-plone-app-event-event-css')